backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.0
pydantic==2.8.2
websockets==12.0
python-dotenv==1.0.1
prometheus-client==0.21.0
numpy==1.26.4
soundfile==0.12.1
pydub==0.25.1
faster-whisper==1.0.0
openai==1.51.0